refactor: add messenger interface
Add a new Messenger interface to the matrix package to make the code
more testable. Since the methods in this package only use two methods on
the matrix client, it makes sense to provide it with an interface that
only has those methods. This makes mocking the object for testing a lot
easier.
Update the rest of the code to use the new interface.