refactor: move forwarding code to grafana package
Move the code that handles forwarding grafana alerts to a matrix Writer
to the `grafana` package. This way the `matrix` package remains focused
on general access to the matrix server, and the `grafana` package is
focused on handling grafana data.
Rename the main method used to send alerts to a matrix room to
`ForwardAlert` to better describe its purpose.