feat: export forward count metrics (#17)
Update the alert forwarding code to keep track of the total number of
alerts forwarded, as well as successful and failed forwards.
Update the metrics endpoint to export the number of forwarded alerts
as well as the "up" metric. The different forwarding states are
exported as labels on the main `forwards` metric.
Add new unit tests for the metrics endpoint.