ci: build and deploy docker image
Update the makefile to inculde a command to build the project dockerfile
using either the `:latest` tag or a tag based on the git tab.
Update the Gitlab CI pipeline to build the docker image and publish it to
the Gitlab registry.
Refactor the existing build step to cache the build output so that it can
be re-used to build the docker image.