feat: support resolving alerts with replies (#20)
Update the alert handler to store the formatted text that was send with
the original alert in the alert map. This data is used to build the reply
message payload (the reply includes the original message in a blockquote).
Add new message templates for the reply messages.
Add new option for the `resolveMode` CLI parameter to configure the tool
to send a reply to the original alert message when the alert is resolved.
Add a new `Reply` method on the matrix writer to handle sending replies to
a given event ID.