refactor: remove custom message structs
Update the matrix writer code to use the `event.MessageEventContent`
struct when building messages instead of using a custom struct. This
allows the reply and send methods to use shared code to build the
message payload - the reply adds the `RelatesTo` field.
The `Send` and `Reply` methods on the matrix writer now take the plain
and formatted strings instead of a message object.
Update the grafana handler to use the new methods on the matrix writer.
Remove the `matrix_test.go` file since it is no longer relevant.