feat: support saving event map to file (#24)
Add support for reading and writing the alert map data to a file on
startup and on each forwarded alert. This allows the tool to be
restarted without losing the map of grafana alert to element event.
Move the data map between the alert ID and the matrix event ID to be
contained inside the `Forwarder` object instead of being stored in the
general `var` block.
Add a new method to construct a new instance of a `Forwarder`.