feat: option to persist alert map (#30)
Add a new CLI & environment options to allow enabling or disabling the
functionality to persist the alert map between restarts. The option
defaults to `true`.
Rename internal properties and methods in the forwarder code to follow
a consistent naming scheme.
Add new page on the documentation site to describe what the
`grafanaMatrixMap` file is for and why it was created.
Re-write the section on environment variables to use a table.