chore: refactor forwarder to use a custom struct vs raw payload
Instead of passing the raw grafana payload to the forwarder code, the
forwarder now accepts a custom data struct. This allows the grafana
paylaod to change without having any impact on the code converting
that data into a grafana alert.
Update the request handler to convert the raw grafana payload to the
new data structure before calling the forwarder.