chore: move metric collection code to new package
Add a new metrics package under the server package to store all code
related to metric collection. This makes it easier to add metric collection
code in other parts of the code.
Refactor the server struct to include an instance of the metrics collector
to ensure a single instance is in use across the entire application.