ci: add nightly docker tag for unstable changes
Update the Gitlab CI/CD process to only publish new images with the
`:latest` tag when building git tags. Builds on the `main` branch will 
now use the `:nightly` docker tag. This helps prevent users getting
automatically updated to a version of the tool that is broken. More
testing is done before creating a git tag than when merging a change to
the main branch.