refactor: split request handler code
Define a new interface that grafana alert handlers should implement. This
interface takes the request data and returns a list of room IDs and the
alert data.
Add a new `v0` handler function that implements the new handler interface
and deals with everything that is specific to this version of grafana. The
rest of the shared logic remains in the main `server` package.
This makes it easier to implement new handlers for different versions of
grafana without having to change shared logic.