feat: support multiple alerts at once
Update the forwarder code to support multiple alerts at the same time. This is a new feature of Grafana's new unified alerting system - a single webhook can include multiple alerts.
If multiple alerts are found in the same payload, they are all treated as individual alerts.
Update the unified webhook parser to use the ID and status from the individual alerts instead of the main payload.