feat: support bearer header authentication (#36)
Add support for authentication on the forward alert endpoints.
The only supported mechanism right now is bearer token.
The scheme and credentials can be set using CLI flags and environment variables.
If authentication is enabled, any request that fails that validation will return a 401 status and not forward the alert.